Move auth package to feature/auth
